Improve Map bounds hit testing.
First, I moved map dimensions (number of tiles and the size of the tiles) into the map json.  This removes some magic constants in the code.

Next, I created a function in the map called IsOutOfBounds that returns true if any part of the incoming rectangle is out of bounds or in a barrier.

Along the way, I simplified the map hit boxes by no longer making them hit boxes as that functionality wasn't being used currently.  We can always add it back later if we need to draw those on screen.